requirements.txt
fastapi
uvicorn
google-generativeai
twilio
python-dotenv